requirements.txt
python-whois
dnspython
requests
python-nmap
